# docs/openapi.yaml
openapi: 3.0.3
info:
  title: GophKeeper API
  version: 1.0.0
paths:
  /health:
    get:
      responses:
        "200":
          description: Server health.
  /api/v1/auth/register:
    post:
      summary: Register a new user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "201":
          description: User registered.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: User already exists.
  /api/v1/auth/login:
    post:
      summary: Authenticate an existing user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "200":
          description: User authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials.
  /api/v1/records:
    get:
      security:
        - bearerAuth: []
      summary: List non-deleted encrypted records.
      responses:
        "200":
          description: Records.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Record"
  /api/v1/records/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      security:
        - bearerAuth: []
      summary: Get one encrypted record.
      responses:
        "200":
          description: Record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Record"
        "404":
          description: Record not found.
    put:
      security:
        - bearerAuth: []
      summary: Create or update an encrypted record.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Record"
      responses:
        "200":
          description: Saved record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Record"
    delete:
      security:
        - bearerAuth: []
      summary: Delete one record.
      responses:
        "200":
          description: Tombstone record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Record"
  /api/v1/records/sync:
    post:
      security:
        - bearerAuth: []
      summary: Merge client records and return server state.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncRequest"
      responses:
        "200":
          description: Merged server state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    AuthRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [username, token]
      properties:
        username:
          type: string
        token:
          type: string
    Record:
      type: object
      required: [id, type, updated_at]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [login_password, text, binary, bank_card, otp]
        payload:
          type: string
          description: Base64 encoded AES-GCM envelope produced by the client.
        updated_at:
          type: string
          format: date-time
        deleted:
          type: boolean
    SyncRequest:
      type: object
      required: [records]
      properties:
        records:
          type: array
          items:
            $ref: "#/components/schemas/Record"
    SyncResponse:
      type: object
      required: [records]
      properties:
        records:
          type: array
          items:
            $ref: "#/components/schemas/Record"